+++ docs/changes.rst
@@ -24,12 +24,14 @@ Weblate 5.16
 * Validation of VCS settings :ref:`push-changes` has been extended.
 * The default values for :ref:`project-translation_review` and :ref:`project-source_review` can be configured in settings.
 * The :setting:`PRIVATE_COMMIT_EMAIL_TEMPLATE` now supports the ``{user_id}`` and ``{site_title}`` variables.
+* The default value for personal translation memory contribution is now based on the :setting:`DEFAULT_AUTOCLEAN_TM` configuration.
 
 .. rubric:: Bug fixes
 
 * Adding plural strings with singular matching existing string is now prohibited for bilingual translations (see :ref:`bimono`).
 * Automatic :ref:`component-repoweb` URL for common code hosting sites.
 * File formats that only differed in encoding (CSV, GWT Properties, Java Properties, iOS Strings) have been merged into single formats.
+* Improved cache isolation for suggestion checks to avoid interference with the parent unit checks.
 * Gracefully handle invalid check flags in :ref:`custom-checks`.
 
 .. rubric:: Compatibility